refactor(registerDir): use readdirSync and accessSync
Either this, or promisifying this method is required in order to use lab-di
in our projects without the build sometimes breaking.
If we want to make this method async, then we have to wait for it in every
project we use it in, so it should be decided early in order to avoid the neccessity
of any dependant packages' refactor.