refactor(di.feature): add initialize the DI step
In this feature we have no "real" application, here we want to test
this DI library, actually we just initialize the DI from user
perspective. On the other hand this initialization is not a part of
the World, so the initialization is moved into a step.

With this move the registerDir have to be synchronous.